// types/database.generated.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          approval_status: string | null
          avatar_url: string | null
          avg_quiz_score: number | null
          community_id: string | null
          courses_completed: number | null
          created_at: string | null
          description: string | null
          email: string | null
          first_name: string | null
          generation_id: string | null
          growth_community: string | null
          id: string
          last_active_at: string | null
          last_name: string | null
          learning_preferences: Json | null
          lessons_completed: number | null
          middle_name: string | null
          must_change_password: boolean | null
          name: string | null
          notification_preferences: Json | null
          school: string | null
          school_id: number | null
          timezone: string | null
          total_learning_time_seconds: number | null
        }
        Insert: {
          approval_status?: string | null
          avatar_url?: string | null
          avg_quiz_score?: number | null
          community_id?: string | null
          courses_completed?: number | null
          created_at?: string | null
          description?: string | null
          email?: string | null
          first_name?: string | null
          generation_id?: string | null
          growth_community?: string | null
          id?: string
          last_active_at?: string | null
          last_name?: string | null
          learning_preferences?: Json | null
          lessons_completed?: number | null
          middle_name?: string | null
          must_change_password?: boolean | null
          name?: string | null
          notification_preferences?: Json | null
          school?: string | null
          school_id?: number | null
          timezone?: string | null
          total_learning_time_seconds?: number | null
        }
        Update: {
          approval_status?: string | null
          avatar_url?: string | null
          avg_quiz_score?: number | null
          community_id?: string | null
          courses_completed?: number | null
          created_at?: string | null
          description?: string | null
          email?: string | null
          first_name?: string | null
          generation_id?: string | null
          growth_community?: string | null
          id?: string
          last_active_at?: string | null
          last_name?: string | null
          learning_preferences?: Json | null
          lessons_completed?: number | null
          middle_name?: string | null
          must_change_password?: boolean | null
          name?: string | null
          notification_preferences?: Json | null
          school?: string | null
          school_id?: number | null
          timezone?: string | null
          total_learning_time_seconds?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "profiles_community_id_fkey"
            columns: ["community_id"]
            isOneToOne: false
            referencedRelation: "community_progress_report"
            referencedColumns: ["community_id"]
          },
          {
            foreignKeyName: "profiles_community_id_fkey"
            columns: ["community_id"]
            isOneToOne: false
            referencedRelation: "growth_communities"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "profiles_generation_id_fkey"
            columns: ["generation_id"]
            isOneToOne: false
            referencedRelation: "generations"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "profiles_school_id_fkey"
            columns: ["school_id"]
            isOneToOne: false
            referencedRelation: "school_progress_report"
            referencedColumns: ["school_id"]
          },
          {
            foreignKeyName: "profiles_school_id_fkey"
            columns: ["school_id"]
            isOneToOne: false
            referencedRelation: "schools"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      red_escuelas: {
        Row: {
          agregado_por: string
          fecha_agregada: string | null
          id: string
          red_id: string
          school_id: number
        }
        Insert: {
          agregado_por: string
          fecha_agregada?: string | null
          id?: string
          red_id: string
          school_id: number
        }
        Update: {
          agregado_por?: string
          fecha_agregada?: string | null
          id?: string
          red_id?: string
          school_id?: number
        }
        Relationships: [
          {
            foreignKeyName: "red_escuelas_red_id_fkey"
            columns: ["red_id"]
            isOneToOne: false
            referencedRelation: "redes_de_colegios"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "red_escuelas_school_id_fkey"
            columns: ["school_id"]
            isOneToOne: false
            referencedRelation: "school_progress_report"
            referencedColumns: ["school_id"]
          },
          {
            foreignKeyName: "red_escuelas_school_id_fkey"
            columns: ["school_id"]
            isOneToOne: false
            referencedRelation: "schools"
            referencedColumns: ["id"]
          },
        ]
      }
      redes_de_colegios: {
        Row: {
          created_at: string | null
          created_by: string
          descripcion: string | null
          id: string
          last_updated_by: string | null
          nombre: string
          updated_at: string | null
        }
        Insert: {
          created_at?: string | null
          created_by: string
          descripcion?: string | null
          id?: string
          last_updated_by?: string | null
          nombre: string
          updated_at?: string | null
        }
        Update: {
          created_at?: string | null
          created_by?: string
          descripcion?: string | null
          id?: string
          last_updated_by?: string | null
          nombre?: string
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
      schools: {
        Row: {
          cliente_id: string | null
          has_generations: boolean | null
          id: number
          name: string
        }
        Insert: {
          cliente_id?: string | null
          has_generations?: boolean | null
          id?: number
          name: string
        }
        Update: {
          cliente_id?: string | null
          has_generations?: boolean | null
          id?: number
          name?: string
        }
        Relationships: [
          {
            foreignKeyName: "schools_cliente_id_fkey"
            columns: ["cliente_id"]
            isOneToOne: false
            referencedRelation: "clientes"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      user_roles: {
        Row: {
          assigned_at: string | null
          assigned_by: string | null
          community_id: string | null
          created_at: string | null
          feedback_scope: Json | null
          generation_id: string | null
          id: string
          is_active: boolean | null
          reporting_scope: Json | null
          role_type: Database["public"]["Enums"]["user_role_type"]
          school_id: number | null
          user_id: string
        }
        Insert: {
          assigned_at?: string | null
          assigned_by?: string | null
          community_id?: string | null
          created_at?: string | null
          feedback_scope?: Json | null
          generation_id?: string | null
          id?: string
          is_active?: boolean | null
          reporting_scope?: Json | null
          role_type: Database["public"]["Enums"]["user_role_type"]
          school_id?: number | null
          user_id: string
        }
        Update: {
          assigned_at?: string | null
          assigned_by?: string | null
          community_id?: string | null
          created_at?: string | null
          feedback_scope?: Json | null
          generation_id?: string | null
          id?: string
          is_active?: boolean | null
          reporting_scope?: Json | null
          role_type?: Database["public"]["Enums"]["user_role_type"]
          school_id?: number | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "user_roles_assigned_by_fkey"
            columns: ["assigned_by"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "user_roles_community_id_fkey"
            columns: ["community_id"]
            isOneToOne: false
            referencedRelation: "community_progress_report"
            referencedColumns: ["community_id"]
          },
          {
            foreignKeyName: "user_roles_community_id_fkey"
            columns: ["community_id"]
            isOneToOne: false
            referencedRelation: "growth_communities"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "user_roles_generation_id_fkey"
            columns: ["generation_id"]
            isOneToOne: false
            referencedRelation: "generations"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "user_roles_school_id_fkey"
            columns: ["school_id"]
            isOneToOne: false
            referencedRelation: "school_progress_report"
            referencedColumns: ["school_id"]
          },
          {
            foreignKeyName: "user_roles_school_id_fkey"
            columns: ["school_id"]
            isOneToOne: false
            referencedRelation: "schools"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "user_roles_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never